pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eegverse"
version = "0.3.0"
description = "EEGVERSE: A Universe of Annotated EEG Data"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your@email.com" }]
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "scipy",
  "pandas",
  "mne>=1.6",
  "mne-bids>=0.13",
  "scikit-learn",
  "torch",
  "pooch",
  "openneuro-py"
]

[project.urls]
Homepage = "https://github.com/yourname/eegverse"
Issues = "https://github.com/yourname/eegverse/issues"
